Wallet

+++ ErfanLearn.Core/DTOs/Users/UserPanelViewModel.cs
@@ -9,7 +9,7 @@ namespace ErfanLearn.Core.DTOs
         public string UserName { get; set; }
         public string Email { get; set; }
         public DateTime RegisterDate { get; set; }
-        public int Wallet { get; set; }
+        public long Wallet { get; set; }
     }
 
     public class SideBarUserPanelViewModel
